build fixes

+++ modules/core/src/main/java/com/ae/apps/lib/custom/views/RoundedImageView.java
@@ -21,7 +21,7 @@ import androidx.appcompat.widget.AppCompatImageView;
  *
  * @author based on the stackoverflow question at
  * <p>
- * http://stackoverflow.com/questions/16208365/create-a-circular-image-view-in-android
+ * <a href="http://stackoverflow.com/questions/16208365/create-a-circular-image-view-in-android">...</a>
  */
 public class RoundedImageView extends AppCompatImageView {
 
@@ -72,7 +72,7 @@ public class RoundedImageView extends AppCompatImageView {
         // getMaxHeight() is min api 16
         // int maxHeight = (getMaxHeight() > 0) ? getMaxHeight() : imageHeight;
 
-        // Check if image height is atleast minheight
+        // Check if image height is at least min height
         if (imageHeight <= minHeight) {
             return minHeight;
         }
@@ -108,8 +108,8 @@ public class RoundedImageView extends AppCompatImageView {
 
         // TODO : remove hardcoding here
         paint.setColor(getResources().getColor(android.R.color.white));
-        float cx = scaledBitmap.getWidth() / 2;
-        float cy = scaledBitmap.getHeight() / 2;
+        float cx = (float) scaledBitmap.getWidth() / 2;
+        float cy = (float) scaledBitmap.getHeight() / 2;
         canvas.drawCircle(cx + 0.7f, cy + 0.7f, cx + 0.1f, paint);
         paint.setXfermode(new PorterDuffXfermode(Mode.SRC_IN));
         canvas.drawBitmap(scaledBitmap, rect, rect, paint);
